+++ tests/test_models.py
@@ -7,7 +7,7 @@ import pytest
 from asdf.exceptions import ValidationError
 from astropy import units as u
 from astropy.modeling import Model
-from astropy.table import QTable
+from astropy.table import QTable, Table
 from numpy.testing import assert_array_equal
 
 from roman_datamodels import datamodels
@@ -754,6 +754,34 @@ def test_make_tvac():
     # Test validation
     tvac_model = datamodels.TvacModel(tvac)
     assert tvac_model.validate() is None
+
+
+def test_make_source_catalog():
+    source_catalog = utils.mk_source_catalog()
+    source_catalog_model = datamodels.SourceCatalogModel(source_catalog)
+
+    assert isinstance(source_catalog_model.source_catalog, Table)
+
+
+def test_make_segmentation_map():
+    segmentation_map = utils.mk_segmentation_map()
+    segmentation_map_model = datamodels.SegmentationMapModel(segmentation_map)
+
+    assert isinstance(segmentation_map_model.data, np.ndarray)
+
+
+def test_make_mosaic_source_catalog():
+    source_catalog = utils.mk_mosaic_source_catalog()
+    source_catalog_model = datamodels.MosaicSourceCatalogModel(source_catalog)
+
+    assert isinstance(source_catalog_model.source_catalog, Table)
+
+
+def test_make_mosaic_segmentation_map():
+    segmentation_map = utils.mk_mosaic_segmentation_map()
+    segmentation_map_model = datamodels.MosaicSegmentationMapModel(segmentation_map)
+
+    assert isinstance(segmentation_map_model.data, np.ndarray)
 
 
 def test_datamodel_info_search(capsys):